requirements.txt
autopep8
Django
django-rest-framework
djangorestframework
drf-yasg

psycopg2-binary
djangorestframework_simplejwt